Lowered minSdkVersion from 15 to 11
Using API level 11 is ok regarding to the API calls. Lower than 11 is not possible because of the ActionBar.

There might be an issue because of the use of "android:fontFamily" in the layout xmls. This is only available from API level 16. But this was existing before already. I think it's negligible.

Fixes #5